Fixed erroneous post-send assertion on the amount of data sent.
The assertion fails if the amount of bytes sent is equal to the payload size. This is wrong because as per the java doc on the `org.apache.qpid.proton.engine
.Sender.send` method these two values will always be equal when successful.

This problem is not usually visible because java assert keyword is not intended for production code and is disabled by default. However, test runners typically enable it by passing the `-ea` flag to the jvm. In my case it was causing the tests to fail.

I’ve also debugged it against the sample code provided with the SDK. Since it uses the static main method the IDE runners do not pass the `-ea` flag and the assertion is not enabled. However, if it were enabled, the send would fail on it despite being successful, i.e., event being successfully written to the event hub.

I suggest releasing this fix under 0.6.5 version.